Working multipart with temp file cleanup
If an exception occurs saving a param to a temp file then clean up that
temp file immediately, and communicate the exception to the caller with
an Either.  Higher level code then looks for `Left`s and if one is found
it first cleans up any previously established temp files and then
rethrows the exception.  It's not that pretty, but it looks to be
working.

One very basic test working.  More challenging will be tests to exercise
the various error conditions.