Handler back to Req => Option[Resp]
I originally had handlers modeled this way.  Why did I change it? Maybe
I'll relearn soon :)

It's "simpler" (syntactically) to not have an optional response, but it
ignores the reality that a handler may not have any idea what to do with
a given request.  This comes up in route matching, for example.  You
have a bunch of routes (which I want to /be/ handlers, for
composability) and you want to try each of them until one returns a
response.  Ring deals with this by just returning nil, but I obviously
want to avoid that in Scala.