Use `mw()` as convention for getting middleware
e.g. Foo.mw() returns a Foo middleware.  `apply` was nice, but it causes
conflicts if you have, say, a case class with the same name as your
middleware, e.g. see Params.scala.

Also moved more stuff to CpsHandler and introduced the ugly
Circle.mockSend continuation fn to force execution of a handler, which
is handy in many cases.